requirements.txt
langchain
langchain_community
python-dotenv
dashscope